refactor: ApiResponse 구조 개선 및 성공/실패 응답 명확화
